+++ datasets/bls/_terraform/cpsaat18_pipeline.tf
@@ -15,20 +15,22 @@
  */
 
 
-resource "google_bigquery_dataset" "{{ dataset_id }}" {
-  dataset_id = "{{ dataset_id }}"
+resource "google_bigquery_table" "cpsaat18" {
   project    = var.project_id
-  {% if friendly_name -%}
-    friendly_name = "{{ friendly_name }}"
-  {% endif -%}
-  {% if description -%}
-    description = {{ description|tojson }}
-  {% endif -%}
-  {% if location -%}
-    location = "{{ location }}"
-  {% endif -%}
+  dataset_id = "bls"
+  table_id   = "cpsaat18"
+
+  description = "Current population survey 18: Employed persons by detailed industry, sex, race, and Hispanic or Latino ethnicity"
+
+  depends_on = [
+    google_bigquery_dataset.bls
+  ]
 }
 
-output "bigquery_dataset-{{ dataset_id }}-dataset_id" {
-  value = google_bigquery_dataset.{{ dataset_id }}.dataset_id
+output "bigquery_table-cpsaat18-table_id" {
+  value = google_bigquery_table.cpsaat18.table_id
+}
+
+output "bigquery_table-cpsaat18-id" {
+  value = google_bigquery_table.cpsaat18.id
 }