+++ datasets/google_trends/_terraform/google_trends_dataset.tf
@@ -15,22 +15,12 @@
  */
 
 
-resource "google_bigquery_dataset" "covid19_tracking" {
-  dataset_id  = "covid19_tracking"
+resource "google_bigquery_dataset" "google_trends" {
+  dataset_id  = "google_trends"
   project     = var.project_id
-  description = "BigQuery dataset for the COVID-19 Tracking Project"
+  description = "The Google Trends dataset will provide critical signals that individual users and businesses alike can leverage to make better data-driven decisions. This dataset simplifies the manual interaction with the existing Google Trends UI by automating and exposing anonymized, aggregated, and indexed search data in BigQuery."
 }
 
-output "bigquery_dataset-covid19_tracking-dataset_id" {
-  value = google_bigquery_dataset.covid19_tracking.dataset_id
-}
-
-resource "google_storage_bucket" "covid-tracking-project" {
-  name                        = "${var.bucket_name_prefix}-covid-tracking-project"
-  force_destroy               = true
-  uniform_bucket_level_access = true
-}
-
-output "storage_bucket-covid-tracking-project-name" {
-  value = google_storage_bucket.covid-tracking-project.name
+output "bigquery_dataset-google_trends-dataset_id" {
+  value = google_bigquery_dataset.google_trends.dataset_id
 }