+++ tests/scripts/test_deploy_dag.py
@@ -175,6 +175,48 @@ def test_script_always_requires_dataset_arg(
 pipeline_path_2 = pipeline_path
 
 
+def test_script_can_deploy_without_variables_files(
+    dataset_path: pathlib.Path,
+    pipeline_path: pathlib.Path,
+    airflow_home: pathlib.Path,
+    env: str,
+    mocker,
+):
+    setup_dag_and_variables(
+        dataset_path,
+        pipeline_path,
+        airflow_home,
+        env,
+        f"{dataset_path.name}_variables.json",
+    )
+
+    # Delete the shared variables file
+    (ENV_DATASETS_PATH / "shared_variables.json").unlink()
+    assert not (ENV_DATASETS_PATH / "shared_variables.json").exists()
+
+    # Delete the dataset-specific variables file
+    (
+        ENV_DATASETS_PATH / dataset_path.name / f"{dataset_path.name}_variables.json"
+    ).unlink()
+    assert not (
+        ENV_DATASETS_PATH / dataset_path.name / f"{dataset_path.name}_variables.json"
+    ).exists()
+
+    mocker.patch("scripts.deploy_dag.run_gsutil_cmd")
+    mocker.patch("scripts.deploy_dag.run_cloud_composer_vars_import")
+
+    deploy_dag.main(
+        local=False,
+        env_path=ENV_PATH,
+        dataset_id=dataset_path.name,
+        pipeline=pipeline_path.name,
+        airflow_home=airflow_home,
+        composer_env="test-env",
+        composer_bucket="test-bucket",
+        composer_region="test-region",
+    )
+
+
 def test_script_with_pipeline_arg_deploys_only_that_pipeline(
     dataset_path: pathlib.Path,
     pipeline_path: pathlib.Path,
@@ -347,6 +389,7 @@ def test_script_with_local_flag_copies_files_to_local_airflow_env(
         composer_region=None,
     )
 
+    assert (airflow_home / "data" / "variables" / "shared_variables.json").exists()
     assert (airflow_home / "data" / "variables" / variables_filename).exists()
     assert (airflow_home / "dags" / dag_filename).exists()